+++ wgpu_shadertoy/__init__.py
@@ -1,4 +1,4 @@
-from .inputs import ShadertoyChannel
+from .inputs import ShadertoyChannel, ShadertoyChannelTexture
 from .shadertoy import Shadertoy
 
 __version__ = "0.1.0"